Add files via upload
In this assignment, we will work with k-nearest neighbors to classify wine quality using pandas, NumPy, and sklearn Python packages. We will use pandas to import the data set and convert the data into NumPy arrays to prepare it for classification using sklearn. We will then split the data set into a training data set and a test data set. From there, we will use sklearn to normalize the data sets, train the classifier, and use cross-validation to select the best k. Finally, we will evaluate the selected classifier using the test data set.